chore: setup Jest, Supertest and refactor app/index separation (#32)
* Phase 3 complete: API Sports client + cron jobs (#25)

* chore: add GitHub issue and PR templates

* chore: initialize Node.js project and install dependencies (#1) (#2)

* feat: create project folder structure and Express base server (#3) (#4)

* chore: add env config and validation (#5) (#6)

* feat: implement error handler middleware (#7) (#8)

* chore: configure MySQL connection pool (#10) (#11)

* db: add initial schema migration for all tables (#12) (#13)

* db: implement all data models with upsert and query methods (#14) (#15)

* chore: implement API client and rate limit guard (#17) (#18)

* chore: implement API-Football response transformers (#19) (#20)

* feat: implement API-Football fetchers for fixtures, standings, players and team stats (#21) (#22)

* feat: register cron jobs for scheduled API fetching (#23) (#24)

* chore: setup Jest, Supertest and refactor app/index separation (#31)
