+++ jest.config.js
@@ -1,0 +1,8 @@
+/** @type {import('jest').Config} */
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['**/tests/**/*.test.js'],
+  clearMocks: true,
+  coverageDirectory: 'coverage',
+  collectCoverageFrom: ['src/**/*.js'],
+};
