+++ src/app.js
@@ -1,0 +1,16 @@
+const express = require('express');
+const cors = require('cors');
+const morgan = require('morgan');
+const routes = require('./routes/index');
+const errorHandler = require('./middlewares/errorHandler');
+
+const app = express();
+
+app.use(cors());
+app.use(morgan('dev'));
+app.use(express.json());
+
+app.use('/api', routes);
+app.use(errorHandler);
+
+module.exports = app;
