+++ src/index.js
@@ -1,24 +1,12 @@
 require('dotenv').config();
-const express = require('express');
-const cors = require('cors');
-const morgan = require('morgan');
 const { validateEnv } = require('./config/env');
 const { testConnection } = require('./config/db');
 const { registerJobs } = require('./jobs/index');
-const routes = require('./routes/index');
-const errorHandler = require('./middlewares/errorHandler');
+const app = require('./app');
 
 validateEnv();
 
-const app = express();
 const PORT = process.env.PORT || 3000;
-
-app.use(cors());
-app.use(morgan('dev'));
-app.use(express.json());
-
-app.use('/api', routes);
-app.use(errorHandler);
 
 const start = async () => {
   await testConnection();
@@ -34,5 +22,3 @@ start().catch((err) => {
   console.error('Failed to start server:', err.message);
   process.exit(1);
 });
-
-module.exports = app;
